Move connection drop to new Op system
